requirements.txt
meshtastic==2.7.7
flask==3.0.3
pypubsub==4.0.3
pyserial==3.5
bleak>=0.21.0
bluez-peripheral==0.1.7
paho-mqtt>=2.0.0
